refactor: moved queries script outside of tdp cli to the root of tdp
