+++ tdp/cli/commands/deploy.py
@@ -8,10 +8,6 @@ from typing import TYPE_CHECKING
 
 import click
 
-from tdp.cli.queries import (
-    get_planned_deployment,
-    get_sch_status,
-)
 from tdp.cli.session import get_session
 from tdp.cli.utils import (
     check_services_cleanliness,
@@ -24,6 +20,10 @@ from tdp.core.cluster_status import ClusterStatus
 from tdp.core.deployment import DeploymentRunner, Executor
 from tdp.core.models import DeploymentStateEnum
 from tdp.core.variables import ClusterVariables
+from tdp.queries import (
+    get_planned_deployment,
+    get_sch_status,
+)
 
 if TYPE_CHECKING:
     from tdp.core.collections import Collections
