+++ tdp/cli/commands/plan/reconfigure.py
@@ -7,7 +7,6 @@ from typing import TYPE_CHECKING, Optional
 
 import click
 
-from tdp.cli.queries import get_planned_deployment, get_sch_status
 from tdp.cli.session import get_session
 from tdp.cli.utils import (
     collections,
@@ -18,6 +17,7 @@ from tdp.cli.utils import (
 )
 from tdp.core.cluster_status import ClusterStatus
 from tdp.core.models import DeploymentModel
+from tdp.queries import get_planned_deployment, get_sch_status
 
 if TYPE_CHECKING:
     from tdp.core.collections import Collections
